US237 Task289
This commit satisfies Task 289 of US 237, adding a button for full
screen mode that allows the model to be shown in full screen mode. This
commit also creates a folder system for easier understanding of what
image goes to which model or page. This commit also rearranges code and
fixes commenting.