Task307 Create. a star field in three.js
Instead of randomly populating a square and hoping none over lapped the 3D object, I used two spheres to set the minimum star distance from the center (1000) and the maximum star distance from the center (20000).
They are so small that when the stars move over pixels its almost as if they are already twinkling. Animating them got me only so far as to change the opaqueness of them all at once which made them look like they were on a single dimmer. It was dumb.
I played around way too much with the size of the stars and the different inner and outer position limits, but these seem to be perfect making some stars appear bigger than others and got rid of them all being roughly the same size.